pyproject.toml
[project]
name = "skrift"
version = "0.1.0a18"
description = "A lightweight async Python CMS for crafting modern websites"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litestar[standard,jinja,cryptography]>=2.14.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "aiosqlite>=0.20.0",
    "asyncpg>=0.30.0",
    "advanced-alchemy>=0.26.0",
    "alembic>=1.14.0",
    "httpx>=0.28.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.34.0",
    "pyyaml>=6.0.0",
    "ruamel.yaml>=0.18.0",
    "markdown-it-py>=3.0.0",
    "mdit-py-plugins>=0.4.0",
]

[project.scripts]
skrift = "skrift.cli:cli"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
docs = [
    "zensical>=0.0.19",
]

[tool.hatch.build]
include = [
    "skrift/**/*.py",
    "skrift/templates/**/*.html",
    "skrift/static/**/*",
    "skrift/alembic/**/*",
    "skrift/alembic.ini",
    "skrift/claude_skill/*.md",
]
exclude = [
    "**/__pycache__",
    "**/*.pyc",
    "app.yaml",
    "*.db",
]

[project.optional-dependencies]
docs = ["zensical>=0.0.19"]

[tool.ruff]
line-length = 100
target-version = "py313"
